chore(security): rename namespace
- From Token into Authentication.
