+++ src/Foundation/Security/Authentication/RefreshTokenManagerInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Foundation\Security\Token;
+namespace App\Foundation\Security\Authentication;
 
 use Gesdinet\JWTRefreshTokenBundle\Model\RefreshTokenInterface as RefreshToken;
 use Symfony\Component\Security\Core\User\UserInterface;
