+++ src/Foundation/Security/Authentication/TokenManagerInterface.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace App\Foundation\Security\Token;
+namespace App\Foundation\Security\Authentication;
 
 use Gesdinet\JWTRefreshTokenBundle\Model\RefreshTokenInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 
 /**
- * Token manager interface.
+ * Authentication manager interface.
  * All access token managers must implement this interface.
  */
 interface TokenManagerInterface
